+++ bin/compile_dependencies.sh
@@ -1,4 +1,15 @@
 #!/bin/bash
+#
+# Compile the dependencies for production, CI and development.
+#
+# Usage, in the root of the project:
+#
+#     ./bin/compile_dependencies.sh
+#
+# Any extra flags/arguments passed to this wrapper script are passed down to pip-compile.
+# E.g. to update a package:
+#
+#     ./bin/compile_dependencies.sh --upgrade-package django
 
 set -ex
 
@@ -6,21 +17,40 @@ toplevel=$(git rev-parse --show-toplevel)
 
 cd $toplevel
 
-# Base (& prod) deps
+export CUSTOM_COMPILE_COMMAND="./bin/compile_dependencies.sh"
+
+# Base deps
 pip-compile \
-    --no-index \
+    --no-emit-index-url \
+    --allow-unsafe \
+    "$@" \
     requirements/base.in
 
-# Dependencies for testing
+# Production deps
 pip-compile \
-    --no-index \
-    --output-file requirements/ci.txt \
+    --no-emit-index-url \
+    --allow-unsafe \
+    --output-file requirements/production.txt \
+    "$@" \
     requirements/base.txt \
-    requirements/test-tools.in
+    requirements/production.in
 
-# Dev depedencies - exact same set as CI + some extra tooling
+# Dev deps
 pip-compile \
-    --no-index \
+    --no-emit-index-url \
+    --allow-unsafe \
     --output-file requirements/dev.txt \
-    requirements/ci.txt \
+    "$@" \
+    requirements/base.txt \
+    requirements/test-tools.in \
     requirements/dev.in
+
+# CI deps
+pip-compile \
+    --no-emit-index-url \
+    --allow-unsafe \
+    --output-file requirements/ci.txt \
+    "$@" \
+    requirements/base.txt \
+    requirements/test-tools.in \
+    requirements/ci.in