+++ src/token_issuer/conf/dev.py
@@ -7,8 +7,6 @@ os.environ.setdefault(
     "SECRET_KEY", "%cgu9q!-d91jh+6+dpx#e2r14c57%ud0o0gm1uttup9rq0f8(4"
 )
 
-os.environ.setdefault("IS_HTTPS", "no")
-
 os.environ.setdefault("DB_NAME", "token_issuer")
 os.environ.setdefault("DB_USER", "token_issuer")
 os.environ.setdefault("DB_PASSWORD", "token_issuer")
@@ -22,8 +20,16 @@ EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
 
 LOGGING["loggers"].update(
     {
-        "token_issuer": {"handlers": ["console"], "level": "DEBUG", "propagate": True,},
-        "django": {"handlers": ["console"], "level": "DEBUG", "propagate": True,},
+        "token_issuer": {
+            "handlers": ["console"],
+            "level": "DEBUG",
+            "propagate": True,
+        },
+        "django": {
+            "handlers": ["console"],
+            "level": "DEBUG",
+            "propagate": True,
+        },
         "django.db.backends": {
             "handlers": ["django"],
             "level": "DEBUG",
@@ -34,9 +40,18 @@ LOGGING["loggers"].update(
             "level": "WARNING",
             "propagate": False,
         },
-        "performance": {"handlers": ["console"], "level": "INFO", "propagate": True,},
+        "performance": {
+            "handlers": ["console"],
+            "level": "INFO",
+            "propagate": True,
+        },
     }
 )
+
+# Disable security measures for development
+SESSION_COOKIE_SECURE = False
+SESSION_COOKIE_HTTPONLY = False
+CSRF_COOKIE_SECURE = False
 
 #
 # Custom settings