+++ src/token_issuer/services/migrations/0002_auto_20181115_2138.py
@@ -23,7 +23,9 @@ class Migration(migrations.Migration):
                     ),
                 ),
             ],
-            options={"verbose_name": "Service configuration",},
+            options={
+                "verbose_name": "Service configuration",
+            },
         ),
         migrations.AddField(
             model_name="service",