+++ src/token_issuer/services/migrations/0004_auto_20190515_1755.py
@@ -12,7 +12,10 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrations.RemoveField(model_name="configuration", name="ztcs",),
+        migrations.RemoveField(
+            model_name="configuration",
+            name="ztcs",
+        ),
         migrations.AddField(
             model_name="configuration",
             name="primary_ac",