+++ src/token_issuer/services/models.py
@@ -1,7 +1,7 @@
 from urllib.parse import urljoin
 
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 import requests
 from solo.models import SingletonModel