+++ CHANGELOG.rst
@@ -2,6 +2,22 @@
 Change history
 ==============
 
+0.2.0-rc1
+=====
+
+*<month, day year>*
+
+* Replace Gulp with Webpack
+* Update Django from ``2.2`` to ``3.2``
+* Require Postgres version ``10`` or above
+* Replace Travis with Github Actions
+* Add ``compile_dependencies`` script to (re)generate python dependencies
+* Apply admin styling from default project
+* Update javascript testing configuration
+* Update ZGW-consumers from ``0.4.1`` to ``0.19.0``
+  * Requires existing services to set its ``Authorization type`` value to ``zgw``
+    to follow previous behavior whenever a client-id and secret were provided
+
 
 0.1.0
 =====